deployment/requirements.txt
requests
numpy
pandas
Pillow
streamlit
keras